tests: split test_merge_prs_queue.py into pmp and any files
Split the 1,771-line monolith (177 test functions) along the natural
fixture boundary: pmp_recipe tests into test_merge_prs_queue_pmp.py
(59 functions) and impl/remed/impl_groups/any_recipe tests into
test_merge_prs_queue_any.py (118 functions). Original file deleted.
